# app/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "P12-Environmental API",
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "http://localhost:8080"
		}
	],
	"paths": {
		"/api/team": {
			"get": {
				"summary": "Retrieve team information",
				"description": "Fetch the JSON file containing information about the team.",
				"responses": {
					"200": {
						"description": "Team information retrieved successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"example": {
										"team": [
											{
												"name": "Alice",
												"role": "Developer"
											},
											{
												"name": "Bob",
												"role": "Designer"
											}
										]
									}
								}
							}
						}
					},
					"404": {
						"description": "File not found."
					}
				}
			}
		},
		"/api/mailing/subscribe": {
			"post": {
				"summary": "Subscribe to the mailing list",
				"description": "Add an email address to the mailing list and send a confirmation email.",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"format": "email"
									}
								},
								"required": [
									"email"
								]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Confirmation email sent successfully."
					},
					"400": {
						"description": "Invalid input or email already subscribed."
					},
					"513": {
						"description": "Failed to send confirmation email due to an internal server error."
					}
				}
			}
		},
		"/api/mailing/unsubscribe": {
			"post": {
				"summary": "Unsubscribe from the mailing list",
				"description": "Remove an email address from the mailing list using the provided email and unsubscribe token.",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"format": "email"
									},
									"unsubscribeToken": {
										"type": "string"
									}
								},
								"required": [
									"email",
									"unsubscribeToken"
								]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successfully unsubscribed."
					},
					"400": {
						"description": "Invalid input provided."
					},
					"404": {
						"description": "Subscription not found or already unsubscribed."
					}
				}
			}
		},
		"/api/contact": {
			"post": {
				"summary": "Submit a contact message",
				"description": "Submit a contact message with the sender's information and message content.",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"firstName": {
										"type": "string"
									},
									"lastName": {
										"type": "string"
									},
									"email": {
										"type": "string",
										"format": "email"
									},
									"message": {
										"type": "string",
										"maxLength": 300
									}
								},
								"required": [
									"firstName",
									"email",
									"message"
								]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Message submitted successfully."
					},
					"400": {
						"description": "Invalid input provided."
					}
				}
			}
		}
	}
}